split up composite type generation
